Added missing escaping of dashes in regexp matching lists (square brackets)
The dash in the SCHEME regexp makes a range from plus until dot, with the side-effect of adding a comma, which I don't think is legal in scheme names. Escaping the dash makes the matching list only have the characters plus, dash, and dot.
The dash in the UNRESERVED regexp has no effect since the dash is at the beginning, but escaping is used elsewhere, so this change is for consistent regexp syntax.