Added missing escaping of dash character in matching lists (square brackets)
The dash in the SCHEME pattern matching list made the list into a range expression from plus to dot, with the side-effect of adding a comma, which I don't think is a legal character in scheme names. Escaping the dash makes the matching list only include separate characters, and no ranges.
The dash in the UNRESERVED pattern had no side-effect as it is at the beginning of the matching list, but since dashes are escaped elsewhere, escaping it here too makes the usage consistent.