Added trimming of directive and policy string
When used to parse an existing policy, I got an UnrecognizedValueItem for a directive that ended with a space, where the real values were parsed correctly, but there was one extra value created, probably the empty string after the last space. Even though this might be a non-conforming policy, I feel this is a bit too strict, or less forgiving, so I've added trimming of both the policy and the directives, so that any space before a semi-colon, or the end of the policy, are removed.
Since the error message mentions lack of trimming, I removed that part of the error message. Why the error was not triggered, I don't know.
(Note: the problematic space was at the end of the policy, but that was not a problem since spaces are only used for value separators).